Added the function
I Implemented a timer in the TimerScreen composable.
The timer is set based on user input for hours and minutes.
When the timer ends, it toggles the status of the 'silence' function in the app. If 'silence' is enabled, it will be disabled. When the 'silence' is disabled, it will be enabled.